Code splitting
Moved button to ./assets
